Add in-progress GameMap interfaces conversion
