+++ curriculum/part-2/10.md
@@ -1,0 +1,44 @@
+# Rendering with the GameMap
+We must now return to our **render.go** file and make some changes so we can actually see our GameMap!
+
+## Code Analysis
+```
+import (
+	"github.com/kettek/goro"
+
+	"myproject/interfaces"
++  "myproject/mapping"
+)
+```
+Here we have added our mapping package to our render file.
+
+---
+```
+-// DrawAll draws all entities to the screen and flushes it.
+-func DrawAll(screen *goro.Screen, entities []interfaces.Entity) {
++// DrawAll draws all entities and the gameMap to the screen and flushes it.
++func DrawAll(screen *goro.Screen, entities []interfaces.Entity, gameMap interfaces.GameMap, colors map[string]goro.Color) {
++  // Draw all the tiles within the game map.
++  for x := 0; x < gameMap.Width(); x++ {
++    for y := 0; y < gameMap.Height(); y++ {
++      if gameMap.IsBlocked(x, y) {
++        screen.SetBackground(x, y, colors["darkWall"])
++      } else {
++        screen.SetBackground(x, y, colors["darkGround"])
++      }
++    }
++  }
++	// Draw all the entities.
+	for _, entity := range entities {
+		DrawEntity(screen, entity)
+  }
+```
+Here we have a fairly large addition that iterates from 0 to the GameMap's width, then from 0 to the GameMap's height within that. From there, we choose to either draw the "darkWall" or the "darkGround" background color if the target tile is reported as blocking movement.
+
+With these changes in place, we can now rebuild the project and you should see much more color and be able to be blocked by the tiles rendered as walls.
+
+## Topics Covered
+Here we have increased the functionality of our `DrawAll` function to allow rendering the game map.
+
+## What's Next
+With this, we are done with part 2. Part 3 will consist of randomly generating a dungeon.
