+++ curriculum/part-2/8.md
@@ -1,65 +1,80 @@
-# Using our GameMap
-Now we can move on to updating our main program to use our GameMap. Please note that we have yet to update our rendering to reflect the changes here, so rebuilding should be reserved for the end of the next lesson.
+# Mapping: The GameMap
+With our Tile type now in place, we can create a new file in our **mapping** directory called **gamemap.go**. This GameMap will hold the map structure and important functions for map generation.
 
 ## Code Analysis
 
 ```
-import (
-	"github.com/kettek/goro"
-	"log"
-
-	"myproject/entity"
-	"myproject/interfaces"
-+	"myproject/mapping"
-)
++package mapping
 ```
-Here we update our imports to include our new mapping package that GameMap resides within. Ensure that you replace **myproject** with whatever you named your module in [../part-1/1](Part 1:1).
+Here we are declaring our GameMap type to exist in our mapping package.
 
 ---
 ```
-		// Our initial variables.
-+		mapWidth, mapHeight := 80, 24
-+
-+		colors := map[string]goro.Color{
-+			"darkWall":   goro.Color{R: 0, G: 0, B: 100, A: 255},
-+			"darkGround": goro.Color{R: 50, G: 50, B: 150, A: 255},
-+		}
-+
-+		gameMap := mapping.GameMap{
-+			Width:  mapWidth,
-+			Height: mapHeight,
-+		}
-+
-+		gameMap.Initialize()
-+
-		player := entity.NewEntity(screen.Columns/2, screen.Rows/2, '@', goro.Style{Foreground: goro.ColorWhite})
-
++// GameMap is our map type for holding our tiles and dimensions.
++type GameMap struct {
++	width, height int
++	tiles         [][]Tile
++}
 ```
-Here we are adding quite a few things. The first is defining `mapWidth` and `mapHeight` variables. We are then defining a string [map](https://tour.golang.org/moretypes/19) to map color names to `goro.Color` types. After this, we create a `gameMap` variable with its Width and Height set to our earlier mapWidth and mapHeight variables. After this, we then call the gameMap's `Initialize()` method to ensure all data structures are set up properly.
+This is our `GameMap` type that contains the map's width, height, and two-dimensional slice of Tile types.
 
 ---
 ```
-			// Draw screen.
--			DrawAll(screen, entities)
-+			DrawAll(screen, entities, gameMap, colors)
-			ClearAll(screen, entities)
++// NewGameMap initializes a GameMap's tiles to match the provided width and height and sets up a few tiles to block movement and sight. Returns a GameMap interface.
++func NewGameMap(width, height int) interfaces.GameMap {
++	g := &GameMap{
++		width:  width,
++		height: height,
++	}
++	g.tiles = make([][]Tile, g.width)
++
++	for x := range g.tiles {
++		g.tiles[x] = make([]Tile, g.height)
++	}
 ```
-Here we are updating our `DrawAll` function to reflect a change we have yet to make. This change passes both the gameMap and our map of colors we created earlier.
+Here is a helper function that we use to create a GameMap, set up its two-dimensional slice of Tile types, and return it as an interface. We are first using golang's built-in [https://tour.golang.org/moretypes/13](make(...)) function to create a slice of slices of Tiles up to the map's width. After this we iterate over each column of `g.Tiles` and create a slice of Tiles within each up to the map's height.
 
 ---
 ```
-				case ActionMove:
--					player.Move(action.X, action.Y)
-+					if !gameMap.IsBlocked(player.X+action.X, player.Y+action.Y) {
-+						player.Move(action.X, action.Y)
-+					}
++	g.tiles[30][22].Flags = BlockMovement | BlockSight
++	g.tiles[31][22].Flags = BlockMovement | BlockSight
++	g.tiles[32][22].Flags = BlockMovement | BlockSight
++
++	return g
++}
 ```
-The final change we are making is to use the `GameMap.IsBlocked(...)` method we defined in the previous lesson. This will disallow the player from moving into an area that is considered to block movement.
+Now we are setting three tiles at **30x22**, **31x22**, and **32x22** to both block movement and sight. We are using the bitwise **|** operator to join the **BlockMovement** and **BlockSight** flag constants, providing a merged value containing both of their bits that is assigned to the particular Tile's Flags property. We then return our GameMap interface and close out the function.
 
-**NOTE**: If you build at this point, it will fail. The next lesson will bring our rendering up to date with our changes here.
+
+---
+```
++// Width returns our GameMap's width.
++func (g *GameMap) Width() int {
++  return g.width
++}
++
++// Height returns our GameMap's height.
++func (g *GameMap) Height() int {
++  return g.height
++}
+```
+`Width()` and `Height()` are getters for the GameMap's **width** and **height** properties. They simply return the value for either property on GameMap when called.
+
+---
+```
++// IsBlocked returns if the given coordinates are blocking movement.
++func (g *GameMap) IsBlocked(x, y int) bool {
++  // Always block if outside our GameMap's bounds.
++  if x < 0 || x >= g.width || y < 0 || y >= g.height {
++    return true
++  }
++	return g.tiles[y][x].Flags&BlockMovement != 0
++}
+```
+`IsBlocked(...)` simply returns a boolean representing whether or not a given tile blocks movement. We are also checking if the requested coordinates are smaller than 0 or larger than our Width and Height, and if so, to return as if blocking. For the return value, we are using the bitwise **&** operator to effectively multiply the Tile's individual bits with BlockMovement's bits. This effectively filters out all bits that are not contained by BlockMovement. If the resulting value is non-zero, then the Tile does have the BlockMovement bit set.
 
 ## Topics Covered
-We have adjusted our main program to use our GameMap type as well as a new Color map.
+We have created the `GameMap` type that uses our Tile type and will create few dummy tiles for testing.
 
 ## What's Next
-Now we can update our rendering to match this file so we can actually render the GameMap and move our character around within it.
+Now we will move on to integrating the `GameMap` with our main program.
