+++ curriculum/part-3/1.go
@@ -1,23 +1,41 @@
 package mapping
 
-// GameMap is our map data type.
+import (
+	"myproject/interfaces"
+)
+
+// GameMap is our map type for holding our tiles and dimensions.
 type GameMap struct {
-	Width, Height int
-	Tiles         [][]Tile
+	width, height int
+	tiles         [][]Tile
 }
 
-// Initialize initializes a GameMap's Tiles to match its Width and Height. It also sets up some coordinates to block movement and sight.
-func (g *GameMap) Initialize() {
-	g.Tiles = make([][]Tile, g.Width)
+// NewGameMap initializes a GameMap's tiles to match the provided width and height and sets up a few tiles to block movement and sight. Returns a GameMap interface.
+func NewGameMap(width, height int) interfaces.GameMap {
+	g := &GameMap{
+		width:  width,
+		height: height,
+	}
+	g.tiles = make([][]Tile, g.width)
 
-	for x := range g.Tiles {
-		g.Tiles[x] = make([]Tile, g.Height)
-		for y := range g.Tiles[x] {
-			g.Tiles[x][y] = Tile{
-				Flags: BlockSight | BlockMovement,
-			}
+	for x := range g.tiles {
+		g.tiles[x] = make([]Tile, g.height)
+		for y := range g.tiles[x] {
+			g.tiles[x][y].Flags = BlockMovement | BlockSight
 		}
 	}
+
+	return g
+}
+
+// Width returns our GameMap's width.
+func (g *GameMap) Width() int {
+	return g.width
+}
+
+// Height returns our GameMap's height.
+func (g *GameMap) Height() int {
+	return g.height
 }
 
 // IsBlocked returns if the given coordinates are blocking movement.
@@ -26,12 +44,12 @@ func (g *GameMap) IsBlocked(x, y int) bool {
 	if !g.InBounds(x, y) {
 		return true
 	}
-	return g.Tiles[x][y].Flags&BlockMovement != 0
+	return g.tiles[y][x].Flags&BlockMovement != 0
 }
 
 // InBounds returns if the given coordinates are within the map's bounds.
 func (g *GameMap) InBounds(x, y int) bool {
-	if x < 0 || x >= g.Width || y < 0 || y >= g.Height {
+	if x < 0 || x >= g.width || y < 0 || y >= g.height {
 		return false
 	}
 	return true
