+++ curriculum/part-3/1.md
@@ -4,25 +4,16 @@ We are now going to adjust our **mapping/gamemap.go**. We will be adjusting the 
 ## Code Analysis
 
 ```
-	for x := range g.Tiles {
-		g.Tiles[x] = make([]Tile, g.Height)
-+		for y := range g.Tiles[x] {
-+			g.Tiles[x][y] = Tile{
-+				Flags: BlockSight | BlockMovement,
-+			}
+	for x := range g.tiles {
+		g.tiles[x] = make([]Tile, g.height)
++		for y := range g.tiles[x] {
++			g.Tiles[x][y].Flags = BlockMovement | BlockSight
 +		}
 	}
 -
--	g.Tiles[30][22] = Tile{
--		Flags: BlockSight | BlockMovement,
--	}
--	g.Tiles[31][22] = Tile{
--		Flags: BlockSight | BlockMovement,
--	}
--	g.Tiles[32][22] = Tile{
--		Flags: BlockSight | BlockMovement,
--	}
-
+-	g.Tiles[30][22].Flags = BlockSight | BlockMovement
+-	g.Tiles[31][22].Flags = BlockSight | BlockMovement
+-	g.Tiles[32][22].Flags = BlockSight | BlockMovement
 ```
 What we have done here is remove the manual testing tile state and replaced it with a for loop that iterates through all tiles and sets them to be blocked. This means that we will be digging rooms out of our map rather than placing them on it.
 
@@ -30,23 +21,23 @@ What we have done here is remove the manual testing tile state and replaced it w
 // IsBlocked returns if the given coordinates are blocking movement.
 func (g *GameMap) IsBlocked(x, y int) bool {
 	// Always block if outside our GameMap's bounds.
--	if x < 0 || x >= g.Width || y < 0 || y >= g.Height {
+-	if x < 0 || x >= g.width || y < 0 || y >= g.height {
 +	if !g.InBounds(x, y) {
 		return true
 	}
-	return g.Tiles[x][y].Flags&BlockMovement != 0
+	return g.tiles[x][y].Flags&BlockMovement != 0
 }
 ```
 Here we are replacing our lengthier bounds check code with a simpler method that will do the bounds checking for us. This is defined next.
 
 ```
-// InBounds returns if the given coordinates are within the map's bounds.
-func (g *GameMap) InBounds(x, y int) bool {
-	if x < 0 || x >= g.Width || y < 0 || y >= g.Height {
-		return false
-	}
-	return true
-}
++// InBounds returns if the given coordinates are within the map's bounds.
++func (g *GameMap) InBounds(x, y int) bool {
++	if x < 0 || x >= g.width || y < 0 || y >= g.height {
++		return false
++	}
++	return true
++}
 ```
 We are creating a separate `InBounds(...)` method so as to reduce unnecessary code verbosity in the places we will need to do bounds checks.
